feat: wbtc gen lev comp (WIP)
